requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
cached-property==1.5.2
celery==5.0.5
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==36.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2.20
django-celery-beat==2.1.0
django-extensions==3.1.3
django-filebrowser==3.14.1
django-filter==2.4.0
django-grappelli==2.15.1
django-import-export==2.5.0
django-model-utils==4.1.1
django-taggit==1.4.0
django-taggit-labels==0.7.0
django-timezone-field==4.1.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-extensions==0.7.0
et-xmlfile==1.1.0
feedparser==6.0.2
filebrowser==1.0.7
Flask==2.0.0
idna==2.10
importlib-metadata==4.0.1
itsdangerous==2.1.0
Jinja2==3.0.3
kombu==5.1.0
MarkupPy==1.14
MarkupSafe==2.1.0
odfpy==1.4.1
openpyxl==3.0.7
optionaldict==0.1.2
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
typing-extensions==3.10.0.0
urllib3==1.26.4
uWSGI==2.0.19.1
vine==5.0.0
wcwidth==0.2.5
wechatpy==1.8.18
Werkzeug==2.0.3
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.12.0
zipp==3.4.1
